A file existence check was added in the deletion-by-id process to avoid an error

+++ server/src/app/api.js
@@ -231,7 +231,9 @@ exports.deleteById = function(req, res) {
 				writeLog(11, [req.params.id, err.message]);
 				res.sendStatus(500);
 			} else if (query) {
-				fs.unlinkSync(query.path);
+				if (fs.existsSync(query.path)) {
+					fs.unlinkSync(query.path);
+				}
 
 				writeLog(30, req.params.id);
 				res.sendStatus(200);
